refactor: move secrets and metadata plugins to runtime
