+++ resources/secrets/Cargo.toml
@@ -7,6 +7,3 @@ description = "Plugin to for managing secrets on shuttle"
 keywords = ["shuttle-service", "secrets"]
 
 [dependencies]
-async-trait = "0.1.56"
-serde_json = "1"
-shuttle-service = { path = "../../service", version = "0.41.0" }
